tcp: CONFIG_TCP_AUTHOPT at the start of the files
Inline functions that use static keys need to be modified to use
static_key_enabled instead of static_branch_unlikely.

Doing this in the header directly causes an avalanche of changes so only
do use ifdef CONFIG_TCP_AUTHOPT and only for the files for which
CONFIG_TCP_AUTHOPT has relevant changes.

This is still not very granular because the core tcp source files for
which this is defined are large.

Signed-off-by: Leonard Crestez <cdleonard@gmail.com>